fix: InitDataConfig 분리 및 TEST 수정
